backend/requirements.txt
click==8.1.3
colorama==0.4.4
contourpy==1.0.5
Flask==2.2.2
fonttools==4.37.2
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
Werkzeug==2.0.2
gunicorn==20.1.0
beautifulsoup4==4.11.1
lxml==4.9.1
typing_extensions==4.3.0
pandas==1.4.2
numpy==1.23.3
requests==2.28.1
xmltodict==0.13.0
Flask-Login==0.6.2
SQLAlchemy==1.4.41
Flask-SQLAlchemy==2.5.1
bcrypt==4.0.0
Flask-Bcrypt==1.0.1
PyJWT==2.4.0